Clarify that the bypassSlowDivision optimization operates on a single BB [v2]
Summary:
Update some comments to be more explicit.

Change bypassSlowDivision and the functions it calls so that they take
BasicBlock*s and Instruction*s, rather than Function::iterator&s and
BasicBlock::iterator&s.

Change the APIs so that the caller is responsible for updating the
iterator, rather than the callee.  This makes control flow much easier
to follow.

Reviewers: tyler.nowicki, hfinkel

Subscribers: llvm-commits

Differential Revision: http://reviews.llvm.org/D15633