Initialize much of AssemblyWriter lazily.
Summary:
The AssemblyWriter constructor is in the critical path of
Value::operator<<, which in turn is in the critical path of most debug
messages.  It currently scans the entire module multiple times, even
though most of that work is usually unnecessary.  This change makes us
lazier.

With this change, I can compile a big Thrust file with
-debug-only=inline in a few minutes.  Before, I gave up after 30m.

Reviewers: majnemer

Subscribers: llvm-commits

Differential Revision: http://reviews.llvm.org/D17792